extracted optional
