+++ src/util/optional.js
@@ -1,0 +1,18 @@
+const context = require("../context");
+
+const optional = () => {
+  let opt = {};
+
+  const env = context.getEnv();
+  Object.keys(env)
+    .filter((key) => !!env[key])
+    .filter((key) => key.toUpperCase().startsWith("INPUT_SLACK-OPTIONAL-"))
+    .forEach((key) => {
+      const slackKey = key.replace("INPUT_SLACK-OPTIONAL-", "").toLowerCase();
+      opt[slackKey] = env[key];
+    });
+
+  return opt;
+};
+
+module.exports = { optional };
